+++ packages/fuels-programs/src/script_calls.rs
@@ -20,7 +20,7 @@ use crate::{
     call_utils::{generate_contract_inputs, generate_contract_outputs},
     contract::SettableContract,
     logs::{map_revert_error, LogDecoder},
-    ReceiptParser,
+    receipt_parser::ReceiptParser,
 };
 
 #[derive(Debug)]
@@ -212,7 +212,7 @@ where
     /// Create a [`FuelCallResponse`] from call receipts
     pub fn get_response(&self, receipts: Vec<Receipt>) -> Result<FuelCallResponse<D>> {
         let token = ReceiptParser::new(&receipts).parse(None, &D::param_type())?;
-        
+
         Ok(FuelCallResponse::new(
             D::from_token(token)?,
             receipts,
